app-docker/backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-psycopg2==0.47b0
